+++ .eslintrc
@@ -23,7 +23,7 @@
     "max-len": 0
   },
   "parserOptions": {
-    "ecmaVersion": 6
+    "ecmaVersion": 2017
   },
   "env": {
     "node": true,
